FLASH-201 feat: 구현체들 모두 Social Info를 반환하도록 수정
